move jdl-application-definition to built-in-options
