+++ jdl/models/jdl-application-configuration-factory.ts
@@ -23,7 +23,7 @@ import StringJDLApplicationConfigurationOption from './string-jdl-application-co
 import IntegerJDLApplicationConfigurationOption from './integer-jdl-application-configuration-option.js';
 import BooleanJDLApplicationConfigurationOption from './boolean-jdl-application-configuration-option.js';
 import ListJDLApplicationConfigurationOption from './list-jdl-application-configuration-option.js';
-import JDLApplicationDefinition, { JDLApplicationOptionTypeValue } from './jdl-application-definition.js';
+import JDLApplicationDefinition, { JDLApplicationOptionTypeValue } from '../built-in-options/jdl-application-definition.js';
 import JDLApplicationConfigurationOption from './jdl-application-configuration-option.js';
 
 const applicationDefinition = new JDLApplicationDefinition();
